Refactor components and fix imports
