+++ src/app/components/core/RatingCheck.tsx
