+++ src/app/pages/VerifyEntryPage.tsx
@@ -2,7 +2,7 @@
 import { Navbar } from "../components/UI/Navbar"
 // import EntradaAutenticada from '../components/VerificarEntrada/EntradaAutenticada'
 // import EntradaInvalida from '../components/VerificarEntrada/EntradaInvalida'
-import ValidarEntrada from '../components/VerifyEntry/ValidateEntry'
+import ValidarEntrada from "../views/ValidateEntry"
 
 export default function VerifyEntryPage() {
   // const [entradaValida, setEntradaValida] = useState(true)
