+++ src/app/views/LatestEntries.tsx
@@ -1,0 +1,9 @@
+import Card from "../components/UI/Card"
+
+export default function LatestEntries() {
+  return (
+    <>
+      <Card />
+    </>
+  )
+}
